pyproject.toml
[project]
name = "cephix-drp"
version = "0.1.0"
description = "Digital robot prototype with a semantic bus and deterministic kernel."
requires-python = ">=3.11"
dependencies = ["aiohttp>=3.9", "PyYAML>=6.0"]

[project.optional-dependencies]
cli = ["rich>=13.7"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
